Move Config.EnabledKProbes to linux_bpf specific file
It requires `KProbeName` which is only available with the `linux_bpf` build tag.
